+++ azext_iot/commands.py
@@ -20,8 +20,6 @@ def load_command_table(self, _):
         cmd_group.command('invoke-device-method', 'iot_device_method')
         cmd_group.command('invoke-module-method', 'iot_device_module_method')
         cmd_group.command('generate-sas-token', 'iot_get_sas_token')
-        cmd_group.command('apply-configuration', 'iot_edge_set_modules',
-                          deprecate_info='az iot edge set-modules')
         cmd_group.command('monitor-events', 'iot_hub_monitor_events')
         cmd_group.command('monitor-feedback', 'iot_hub_monitor_feedback')
