+++ azext_iot/operations/pnp.py
@@ -12,7 +12,7 @@ from azext_iot.common.shared import SdkType, PnPModelType
 from azext_iot.common._azure import get_iot_pnp_connection_string
 from azext_iot.sdk.pnp.models import SearchOptions
 from azext_iot._factory import _bind_sdk
-from azext_iot._constants import PNP_API_VERSION, PNP_ENDPOINT
+from azext_iot.constants import PNP_API_VERSION, PNP_ENDPOINT
 from azext_iot.common.utility import (unpack_pnp_http_error,
                                       get_sas_token,
                                       shell_safe_json_parse)
