+++ azext_iot/tests/test_iot_digitaltwin_unit.py
@@ -3,7 +3,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-# pylint: disable=W0613,W0621
 
 import pytest
 import json
