+++ azext_iot/tests/test_iot_ext_int.py
@@ -8,20 +8,11 @@ import os
 import random
 import json
 import pytest
-import sys
 
-from uuid import uuid4
-from azure.cli.testsdk import LiveScenarioTest
 from azure.cli.core.util import read_file_content
-from azext_iot.common.utility import (
-    validate_min_python_version,
-    execute_onthread,
-    calculate_millisec_since_unix_epoch_utc,
-)
-from azext_iot._constants import DEVICE_DEVICESCOPE_PREFIX
+from . import IoTLiveScenarioTest
+from azext_iot.constants import DEVICE_DEVICESCOPE_PREFIX
 
-# Add test tools to path
-sys.path.append(os.path.abspath(os.path.join(".", "iotext_test_tools")))
 
 # Set these to the proper IoT Hub, IoT Hub Cstring and Resource Group for Live Integration Tests.
 LIVE_HUB = os.environ.get("azext_iot_testhub")
@@ -44,121 +35,14 @@ CWD = os.path.dirname(os.path.abspath(__file__))
 PRIMARY_THUMBPRINT = "A361EA6A7119A8B0B7BBFFA2EAFDAD1F9D5BED8C"
 SECONDARY_THUMBPRINT = "14963E8F3BA5B3984110B3C1CA8E8B8988599087"
 
-DEVICE_PREFIX = "test-device-"
 
-
-class TestIoTHub(LiveScenarioTest):
-    def __init__(self, _):
-        from . import DummyCliOutputProducer
-
-        super(TestIoTHub, self).__init__(_)
-        self.cli_ctx = DummyCliOutputProducer()
-
-    def setUp(self):
-        self._entity_names = None
-
-    # TODO: @digimaun - Maybe put a helper like this in the shared lib, when you create it?
-    def command_execute_assert(self, command, asserts):
-        from . import capture_output
-
-        with capture_output() as buffer:
-            self.cmd(command, checks=None)
-            output = buffer.get_output()
-
-        for a in asserts:
-            assert a in output
-
-    def _create_entity_names(self, devices=0, edge_devices=0, modules=0, configs=0):
-        result = {}
-        if devices:
-            device_ids = []
-            for _ in range(devices):
-                device_ids.append(
-                    self.create_random_name(prefix=DEVICE_PREFIX, length=32)
-                )
-            result["device_ids"] = device_ids
-
-        if edge_devices:
-            edge_device_ids = []
-            for _ in range(edge_devices):
-                edge_device_ids.append(
-                    self.create_random_name(prefix="test-edge-device-", length=32)
-                )
-            result["edge_device_ids"] = edge_device_ids
-
-        if modules:
-            module_ids = []
-            for _ in range(modules):
-                module_ids.append(
-                    self.create_random_name(prefix="test-module-", length=32)
-                )
-            result["module_ids"] = module_ids
-
-        if configs:
-            config_ids = []
-            for _ in range(configs):
-                config_ids.append(
-                    self.create_random_name(prefix="test-config-", length=32)
-                )
-            result["config_ids"] = config_ids
-
-        self._entity_names = result
-        return result
-
-    def _remove_entities(self, names=None):
-        if not names:
-            names = self._entity_names
-            self._entity_names = None
-
-        device_ids = names.get("device_ids")
-        if not device_ids:
-            device_ids = []
-        edge_device_ids = names.get("edge_device_ids")
-        if edge_device_ids:
-            device_ids.extend(edge_device_ids)
-
-        for i in device_ids:
-            if device_ids.index(i) == (len(device_ids) - 1):
-                # With connection string
-                self.cmd(
-                    "iot hub device-identity delete -d {} --login {}".format(
-                        i, LIVE_HUB_CS
-                    ),
-                    checks=self.is_empty(),
-                )
-            else:
-                self.cmd(
-                    "iot hub device-identity delete -d {} -n {} -g {}".format(
-                        i, LIVE_HUB, LIVE_RG
-                    ),
-                    checks=self.is_empty(),
-                )
-
-        config_ids = names.get("config_ids")
-        if config_ids:
-            for i in config_ids:
-                if config_ids.index(i) == (len(config_ids) - 1):
-                    # With connection string
-                    self.cmd(
-                        "iot hub configuration delete -c {} --login {}".format(
-                            i, LIVE_HUB_CS
-                        ),
-                        checks=self.is_empty(),
-                    )
-                else:
-                    self.cmd(
-                        "iot hub configuration delete -c {} -n {} -g {}".format(
-                            i, LIVE_HUB, LIVE_RG
-                        ),
-                        checks=self.is_empty(),
-                    )
-
-    def tearDown(self):
-        if self._entity_names:
-            self._remove_entities()
+class TestIoTHub(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHub, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
 
     def test_hub(self):
-
         self.cmd(
             "az iot hub generate-sas-token -n {} -g {}".format(LIVE_HUB, LIVE_RG),
             checks=[self.exists("sas")],
@@ -223,15 +107,20 @@ class TestIoTHub(LiveScenarioTest):
             expect_failure=True,
         )
 
+
+class TestIoTHubDevices(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHubDevices, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_hub_devices(self):
         device_count = 5
         edge_device_count = 2
 
-        names = self._create_entity_names(
-            devices=device_count, edge_devices=edge_device_count
-        )
-        device_ids = names["device_ids"]
-        edge_device_ids = names["edge_device_ids"]
+        device_ids = self.generate_device_names(device_count)
+        edge_device_ids = self.generate_device_names(edge_device_count, edge=True)
+        total_devices = device_count + edge_device_count
 
         self.cmd(
             "iot hub device-identity create -d {} -n {} -g {}".format(
@@ -248,13 +137,13 @@ class TestIoTHub(LiveScenarioTest):
             ],
         )
 
-        for i in range(edge_device_count):
+        for edge_device in edge_device_ids:
             self.cmd(
                 "iot hub device-identity create -d {} -n {} -g {} --ee --add-children {} --force".format(
-                    edge_device_ids[i], LIVE_HUB, LIVE_RG, device_ids[4]
+                    edge_device, LIVE_HUB, LIVE_RG, device_ids[4]
                 ),
                 checks=[
-                    self.check("deviceId", edge_device_ids[i]),
+                    self.check("deviceId", edge_device),
                     self.check("status", "enabled"),
                     self.check("statusReason", None),
                     self.check("connectionState", "Disconnected"),
@@ -265,7 +154,7 @@ class TestIoTHub(LiveScenarioTest):
             )
 
             device_scope_str_pattern = r"^{}{}-".format(
-                DEVICE_DEVICESCOPE_PREFIX, edge_device_ids[i]
+                DEVICE_DEVICESCOPE_PREFIX, edge_device
             )
             self.cmd(
                 "iot hub device-identity show -d {} -n {} -g {}".format(
@@ -277,10 +166,11 @@ class TestIoTHub(LiveScenarioTest):
                 ],
             )
 
+        # All edge devices + child device
         query_checks = [self.check("length([*])", edge_device_count + 1)]
-        for i in range(edge_device_count):
+        for i in edge_device_ids:
             query_checks.append(
-                self.exists("[?deviceId==`{}`]".format(edge_device_ids[i]))
+                self.exists("[?deviceId==`{}`]".format(i))
             )
         query_checks.append(self.exists("[?deviceId==`{}`]".format(device_ids[4])))
 
@@ -315,6 +205,7 @@ class TestIoTHub(LiveScenarioTest):
             checks=query_checks,
         )
 
+        # -1 for no return limit
         self.cmd(
             'iot hub query -q "{}" --login {} --top -1'.format(
                 "select * from devices", LIVE_HUB_CS
@@ -444,20 +335,22 @@ class TestIoTHub(LiveScenarioTest):
             ],
         )
 
+        # List all devices
         self.cmd(
             "iot hub device-identity list --hub-name {} --resource-group {}".format(
                 LIVE_HUB, LIVE_RG
             ),
-            checks=[self.check("length([*])", device_count + edge_device_count)],
+            checks=[self.check("length([*])", total_devices)],
         )
 
         self.cmd(
             "iot hub device-identity list --hub-name {} --resource-group {} --top -1".format(
                 LIVE_HUB, LIVE_RG
             ),
-            checks=[self.check("length([*])", device_count + edge_device_count)],
+            checks=[self.check("length([*])", total_devices)],
         )
 
+        # List only edge devices
         self.cmd(
             "iot hub device-identity list -n {} -g {} --ee".format(LIVE_HUB, LIVE_RG),
             checks=[self.check("length([*])", edge_device_count)],
@@ -596,91 +489,26 @@ class TestIoTHub(LiveScenarioTest):
             expect_failure=True,
         )
 
-    def test_hub_apply_device_content(self):
-        edge_device_count = 2
-        names = self._create_entity_names(edge_devices=edge_device_count)
-        edge_device_ids = names["edge_device_ids"]
 
-        for device_id in edge_device_ids:
-            self.cmd(
-                "iot hub device-identity create -d {} -n {} -g {} --ee".format(
-                    device_id, LIVE_HUB, LIVE_RG
-                )
-            )
-
-        content_path = os.path.join(CWD, "test_config_modules_content.json")
-        content_path_malformed = os.path.join(CWD, "test_config_modules_content_malformed.json")
-        content_path_v1 = os.path.join(CWD, "test_config_modules_content_v1.json")
-
-        self.kwargs["generic_content"] = read_file_content(content_path)
-
-        self.cmd(
-            "iot hub apply-configuration -d {} -n {} -g {} -k '{}'".format(
-                edge_device_ids[0], LIVE_HUB, LIVE_RG, content_path
-            ),
-            checks=[self.check("length([*])", 3)],
-        )
-
-        self.cmd(
-            "iot hub apply-configuration -d {} -n {} -g {} --content '{}'".format(
-                edge_device_ids[0], LIVE_HUB, LIVE_RG, "{generic_content}"
-            ),
-            self.check("length([*])", 3),
-        )
-
-        # With connection string
-        self.cmd(
-            "iot hub apply-configuration -d {} --login {} -k '{}'".format(
-                edge_device_ids[0], LIVE_HUB_CS, content_path_v1
-            ),
-            checks=[self.check("length([*])", 4)],
-        )
-
-        # iot edge set-modules replaces apply-configuration
-        self.cmd(
-            "iot edge set-modules -d {} -n {} -g {} -k '{}'".format(
-                edge_device_ids[1], LIVE_HUB, LIVE_RG, content_path
-            ),
-            checks=[self.check("length([*])", 3)],
-        )
-
-        self.cmd(
-            "iot edge set-modules -d {} -n {} -g {} --content '{}'".format(
-                edge_device_ids[1], LIVE_HUB, LIVE_RG, "{generic_content}"
-            ),
-            self.check("length([*])", 3),
-        )
-
-        # With connection string
-        self.cmd(
-            "iot edge set-modules -d {} --login {} -k '{}'".format(
-                edge_device_ids[1], LIVE_HUB_CS, content_path_v1
-            ),
-            checks=[self.check("length([*])", 4)],
-        )
-
-        # Error schema validation - Malformed deployment content causes validation error
-        self.cmd(
-            "iot edge set-modules -d {} -n {} -g {} -k '{}'".format(
-                edge_device_ids[1], LIVE_HUB, LIVE_RG, content_path_malformed
-            ),
-            expect_failure=True,
+class TestIoTHubDeviceTwins(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHubDeviceTwins, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
         )
 
     def test_hub_device_twins(self):
         self.kwargs["generic_dict"] = {"key": "value"}
         self.kwargs["bad_format"] = "{'key: 'value'}"
+
         device_count = 3
+        device_ids = self.generate_device_names(device_count)
 
-        names = self._create_entity_names(devices=device_count)
-        device_ids = names["device_ids"]
-
-        for i in range(device_count):
+        for device in device_ids:
             self.cmd(
                 "iot hub device-identity create -d {} -n {} -g {}".format(
-                    device_ids[i], LIVE_HUB, LIVE_RG
+                    device, LIVE_HUB, LIVE_RG
                 ),
-                checks=[self.check("deviceId", device_ids[i])],
+                checks=[self.check("deviceId", device)],
             )
 
         self.cmd(
@@ -716,6 +544,7 @@ class TestIoTHub(LiveScenarioTest):
         assert result["deviceId"] == device_ids[0]
         assert result["properties"]["desired"]["special"]["key"] == "value"
 
+        # Removal of desired property from twin
         result = self.cmd(
             'iot hub device-twin update -d {} -n {} -g {} --set properties.desired.special="null"'.format(
                 device_ids[0], LIVE_HUB, LIVE_RG
@@ -783,6 +612,7 @@ class TestIoTHub(LiveScenarioTest):
             ],
         )
 
+        # TODO move distributed tracing tests
         self.cmd(
             "iot hub distributed-tracing show -d {} -n {} -g {}".format(
                 device_ids[2], LIVE_HUB, LIVE_RG
@@ -800,17 +630,21 @@ class TestIoTHub(LiveScenarioTest):
         assert result["samplingRate"] == "50%"
         assert not result["isSynced"]
 
+
+class TestIoTHubModules(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHubModules, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_hub_modules(self):
         edge_device_count = 2
         device_count = 1
         module_count = 2
 
-        names = self._create_entity_names(
-            edge_devices=edge_device_count, devices=device_count, modules=module_count
-        )
-        edge_device_ids = names["edge_device_ids"]
-        module_ids = names["module_ids"]
-        device_ids = names["device_ids"]
+        edge_device_ids = self.generate_device_names(edge_device_count, edge=True)
+        device_ids = self.generate_device_names(device_count)
+        module_ids = self.generate_module_names(module_count)
 
         for edge_device in edge_device_ids:
             self.cmd(
@@ -1069,7 +903,7 @@ class TestIoTHub(LiveScenarioTest):
         )
 
         for i in module_ids:
-            if module_ids.index(i) == (len(module_ids) - 1):
+            if module_ids.index(i) == (module_count - 1):
                 # With connection string
                 self.cmd(
                     "iot hub module-identity delete -d {} --login {} --module-id {}".format(
@@ -1085,19 +919,24 @@ class TestIoTHub(LiveScenarioTest):
                     checks=self.is_empty(),
                 )
 
+
+class TestIoTHubModuleTwins(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHubModuleTwins, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_hub_module_twins(self):
         self.kwargs["generic_dict"] = {"key": "value"}
         self.kwargs["bad_format"] = "{'key: 'value'}"
+
         edge_device_count = 1
         device_count = 1
         module_count = 1
 
-        names = self._create_entity_names(
-            edge_devices=edge_device_count, modules=module_count, devices=device_count
-        )
-        edge_device_ids = names["edge_device_ids"]
-        module_ids = names["module_ids"]
-        device_ids = names["device_ids"]
+        edge_device_ids = self.generate_device_names(edge_device_count, True)
+        device_ids = self.generate_device_names(device_count)
+        module_ids = self.generate_module_names(module_count)
 
         self.cmd(
             "iot hub device-identity create -d {} -n {} -g {} --ee".format(
@@ -1253,12 +1092,19 @@ class TestIoTHub(LiveScenarioTest):
                 checks=self.is_empty(),
             )
 
+
+class TestIoTHubDeviceConfigs(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTHubDeviceConfigs, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_device_configurations(self):
         self.kwargs["generic_dict"] = {"key": "value"}
         self.kwargs["bad_format"] = "{'key: 'value'}"
+
         config_count = 5
-        names = self._create_entity_names(configs=config_count)
-        config_ids = names["config_ids"]
+        config_ids = self.generate_config_names(config_count)
 
         content_path = os.path.join(CWD, "test_config_device_content.json")
         metrics_path = os.path.join(CWD, "test_config_device_metrics.json")
@@ -1572,11 +1418,75 @@ class TestIoTHub(LiveScenarioTest):
             expect_failure=True,
         )
 
+
+class TestIoTEdge(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTEdge, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
+    def test_edge_set_modules(self):
+        edge_device_count = 2
+        edge_device_ids = self.generate_device_names(edge_device_count, True)
+
+        for device_id in edge_device_ids:
+            self.cmd(
+                "iot hub device-identity create -d {} -n {} -g {} --ee".format(
+                    device_id, LIVE_HUB, LIVE_RG
+                )
+            )
+
+        content_path = os.path.join(CWD, "test_config_modules_content.json")
+        content_path_malformed = os.path.join(
+            CWD, "test_config_modules_content_malformed.json"
+        )
+        content_path_v1 = os.path.join(CWD, "test_config_modules_content_v1.json")
+
+        self.kwargs["generic_content"] = read_file_content(content_path)
+
+        # iot edge set-modules replaces apply-configuration
+        self.cmd(
+            "iot edge set-modules -d {} -n {} -g {} -k '{}'".format(
+                edge_device_ids[1], LIVE_HUB, LIVE_RG, content_path
+            ),
+            checks=[self.check("length([*])", 3)],
+        )
+
+        self.cmd(
+            "iot edge set-modules -d {} -n {} -g {} --content '{}'".format(
+                edge_device_ids[1], LIVE_HUB, LIVE_RG, "{generic_content}"
+            ),
+            self.check("length([*])", 3),
+        )
+
+        # With connection string
+        self.cmd(
+            "iot edge set-modules -d {} --login {} -k '{}'".format(
+                edge_device_ids[1], LIVE_HUB_CS, content_path_v1
+            ),
+            checks=[self.check("length([*])", 4)],
+        )
+
+        # Error schema validation - Malformed deployment content causes validation error
+        self.cmd(
+            "iot edge set-modules -d {} -n {} -g {} -k '{}'".format(
+                edge_device_ids[1], LIVE_HUB, LIVE_RG, content_path_malformed
+            ),
+            expect_failure=True,
+        )
+
+
+class TestIoTEdgeDeployments(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTEdgeDeployments, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_edge_deployments(self):
         self.kwargs["generic_dict"] = {"key": "value"}
+
         config_count = 3
-        names = self._create_entity_names(configs=config_count)
-        config_ids = names["config_ids"]
+        config_ids = self.generate_config_names(config_count)
 
         content_path = os.path.join(CWD, "test_config_modules_content.json")
         content_path_malformed = os.path.join(
@@ -1812,13 +1722,14 @@ class TestIoTHub(LiveScenarioTest):
                 config_ids[1], LIVE_HUB, LIVE_RG
             )
         )
-        del config_ids[1]
+        del self.config_ids[1]
+
         self.cmd(
             "iot edge deployment delete -d {} --login {}".format(
                 config_ids[0], LIVE_HUB_CS
             )
         )
-        del config_ids[0]
+        del self.config_ids[0]
 
         # Error top of -1 does not work with configurations
         self.cmd(
@@ -1832,608 +1743,11 @@ class TestIoTHub(LiveScenarioTest):
             expect_failure=True,
         )
 
-    @pytest.mark.skipif(
-        not validate_min_python_version(3, 4, exit_on_fail=False),
-        reason="minimum python version not satisfied",
-    )
-    def test_uamqp_device_messaging(self):
-        device_count = 1
 
-        names = self._create_entity_names(devices=device_count)
-        device_ids = names["device_ids"]
-
-        self.cmd(
-            "iot hub device-identity create -d {} -n {} -g {} --ee".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            ),
-            checks=[self.check("deviceId", device_ids[0])],
-        )
-
-        test_body = str(uuid4())
-        test_props = "key0=value0;key1=value1"
-        test_cid = str(uuid4())
-
-        self.cmd(
-            "iot device c2d-message send -d {} --hub-name {} -g {} --data {} --cid {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, test_body, test_cid
-            ),
-            checks=self.is_empty(),
-        )
-
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --hub-name {} -g {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            )
-        ).get_output_in_json()
-
-        assert result["data"] == test_body
-        assert result["correlationId"] == test_cid
-        assert result["ack"] == "none"
-        etag = result["etag"]
-
-        self.cmd(
-            "iot device c2d-message complete -d {} --hub-name {} -g {} --etag {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, etag
-            ),
-            checks=self.is_empty(),
-        )
-
-        test_body = str(uuid4())
-        test_cid = str(uuid4())
-
-        # With connection string
-        self.cmd(
-            "iot device c2d-message send -d {} --data {} --props {} --cid {} --ack {} --login {}".format(
-                device_ids[0], test_body, test_props, test_cid, "positive", LIVE_HUB_CS
-            ),
-            checks=self.is_empty(),
-        )
-
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --login {}".format(
-                device_ids[0], LIVE_HUB_CS
-            )
-        ).get_output_in_json()
-
-        assert result["data"] == test_body
-        assert result["correlationId"] == test_cid
-        assert result["ack"] == "positive"
-        etag = result["etag"]
-
-        self.cmd(
-            "iot device c2d-message reject -d {} --etag {} --login {}".format(
-                device_ids[0], etag, LIVE_HUB_CS
-            ),
-            checks=self.is_empty(),
-        )
-
-        # Test waiting for ack from c2d send
-        from azext_iot.operations.hub import iot_simulate_device
-        from azext_iot._factory import iot_hub_service_factory
-        from azure.cli.core.mock import DummyCli
-
-        cli_ctx = DummyCli()
-        client = iot_hub_service_factory(cli_ctx)
-
-        token, thread = execute_onthread(
-            method=iot_simulate_device,
-            args=[
-                client,
-                device_ids[0],
-                LIVE_HUB,
-                "complete",
-                "Ping from c2d ack wait test",
-                2,
-                5,
-                "http",
-            ],
-            max_runs=5,
-            return_handle=True,
-        )
-
-        self.cmd(
-            "iot device c2d-message send -d {} --ack {} --login {} --wait -y".format(
-                device_ids[0], "full", LIVE_HUB_CS
-            )
-        )
-        token.set()
-        thread.join()
-
-        # Error - invalid wait when no ack requested
-        self.cmd(
-            "iot device c2d-message send -d {} --login {} --wait -y".format(
-                device_ids[0], LIVE_HUB_CS
-            ),
-            expect_failure=True,
-        )
-
-    def test_device_messaging(self):
-        device_count = 1
-
-        names = self._create_entity_names(devices=device_count)
-        device_ids = names["device_ids"]
-
-        self.cmd(
-            "iot hub device-identity create -d {} -n {} -g {} --ee".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            ),
-            checks=[self.check("deviceId", device_ids[0])],
-        )
-
-        self.cmd(
-            "iot device c2d-message receive -d {} --hub-name {} -g {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            ),
-            checks=self.is_empty(),
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device c2d-message receive -d {} --login {}".format(
-                device_ids[0], LIVE_HUB_CS
-            ),
-            checks=self.is_empty(),
-        )
-
-        etag = "00000000-0000-0000-0000-000000000000"
-        self.cmd(
-            "iot device c2d-message complete -d {} --hub-name {} -g {} -e {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, etag
-            ),
-            expect_failure=True,
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device c2d-message complete -d {} --login {} -e {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            ),
-            expect_failure=True,
-        )
-
-        self.cmd(
-            "iot device c2d-message reject -d {} --hub-name {} -g {} -e {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, etag
-            ),
-            expect_failure=True,
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device c2d-message reject -d {} --login {} -e {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            ),
-            expect_failure=True,
-        )
-
-        self.cmd(
-            "iot device c2d-message abandon -d {} --hub-name {} -g {} --etag {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, etag
-            ),
-            expect_failure=True,
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device c2d-message abandon -d {} --login {} --etag {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            ),
-            expect_failure=True,
-        )
-
-        self.cmd(
-            "iot device simulate -d {} -n {} -g {} --mc {} --mi {} --data '{}' --rs 'complete'".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, 2, 1, "IoT Ext Test"
-            ),
-            checks=self.is_empty(),
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device simulate -d {} --login {} --mc {} --mi {} --data '{}' --rs 'complete'".format(
-                device_ids[0], LIVE_HUB_CS, 2, 1, "IoT Ext Test"
-            ),
-            checks=self.is_empty(),
-        )
-
-        self.cmd(
-            "iot device simulate -d {} -n {} -g {} --mc {} --mi {} --data '{}' --rs 'abandon' --protocol http".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, 2, 1, "IoT Ext Test"
-            ),
-            checks=self.is_empty(),
-        )
-
-        # With connection string
-        self.cmd(
-            "iot device simulate -d {} --login {} --mc {} --mi {} --data '{}' --rs 'abandon' --protocol http".format(
-                device_ids[0], LIVE_HUB_CS, 2, 1, "IoT Ext Test"
-            ),
-            checks=self.is_empty(),
-        )
-
-        self.cmd(
-            "iot device simulate -d {} -n {} -g {} --data '{}' --rs 'reject'".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, "IoT Ext Test"
-            ),
-            checks=self.is_empty(),
-            expect_failure=True,
-        )
-
-        self.cmd(
-            "iot device send-d2c-message -d {} -n {} -g {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            ),
-            checks=self.is_empty(),
-        )
-
-        self.cmd(
-            'iot device send-d2c-message -d {} -n {} -g {} --props "MessageId=12345;CorrelationId=54321"'.format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            ),
-            checks=self.is_empty(),
-        )
-
-        # With connection string
-        self.cmd(
-            'iot device send-d2c-message -d {} --login {} --props "MessageId=12345;CorrelationId=54321"'.format(
-                device_ids[0], LIVE_HUB_CS
-            ),
-            checks=self.is_empty(),
-        )
-
-    @pytest.mark.skipif(
-        not validate_min_python_version(3, 5, exit_on_fail=False),
-        reason="minimum python version not satisfied",
-    )
-    def test_hub_monitor_events(self):
-        for cg in LIVE_CONSUMER_GROUPS:
-            self.cmd(
-                "az iot hub consumer-group create --hub-name {} --resource-group {} --name {}".format(
-                    LIVE_HUB, LIVE_RG, cg
-                ),
-                checks=[self.check("name", cg)],
-            )
-
-        from azext_iot.operations.hub import iot_device_send_message
-        from azext_iot._factory import iot_hub_service_factory
-        from azure.cli.core.mock import DummyCli
-
-        cli_ctx = DummyCli()
-        client = iot_hub_service_factory(cli_ctx)
-        device_count = 10
-
-        # Test with invalid connection string
-        self.cmd(
-            "iot hub monitor-events -t 1 -y --login {}".format(LIVE_HUB_CS + "zzz"),
-            expect_failure=True,
-        )
-
-        # Create and Simulate Devices
-        device_ids = self._create_entity_names(devices=device_count)["device_ids"]
-
-        for i in range(device_count):
-            self.cmd(
-                "iot hub device-identity create -d {} -n {} -g {}".format(
-                    device_ids[i], LIVE_HUB, LIVE_RG
-                ),
-                checks=[self.check("deviceId", device_ids[i])],
-            )
-
-        enqueued_time = calculate_millisec_since_unix_epoch_utc()
-
-        for i in range(device_count):
-            execute_onthread(
-                method=iot_device_send_message,
-                args=[
-                    client,
-                    device_ids[i],
-                    LIVE_HUB,
-                    '{\r\n"payload_data1":"payload_value1"\r\n}',
-                    "$.mid=12345;key0=value0;key1=1",
-                    1,
-                    LIVE_RG,
-                ],
-                max_runs=1,
-            )
-        # Monitor events for all devices and include sys, anno, app
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 -y -p sys anno app".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[0], enqueued_time
-            ),
-            device_ids + [
-                "system",
-                "annotations",
-                "application",
-                '"message_id": "12345"',
-                '"key0": "value0"',
-                '"key1": "1"',
-            ],
-        )
-
-        # Monitor events for a single device
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} -d {} --cg {} --et {} -t 10 -y -p sys anno app".format(
-                LIVE_HUB, LIVE_RG, device_ids[0], LIVE_CONSUMER_GROUPS[1], enqueued_time
-            ),
-            [
-                device_ids[0],
-                "system",
-                "annotations",
-                "application",
-                '"message_id": "12345"',
-                '"key0": "value0"',
-                '"key1": "1"',
-            ],
-        )
-
-        # Monitor events with device-id wildcards
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} -d {} --et {} -t 10 -y -p sys anno app".format(
-                LIVE_HUB, LIVE_RG, DEVICE_PREFIX + "*", enqueued_time
-            ),
-            device_ids,
-        )
-
-        # Monitor events for specific devices using query language
-        device_subset_include = device_ids[: device_count // 2]
-        device_include_string = ", ".join(
-            ["'" + deviceId + "'" for deviceId in device_subset_include]
-        )
-        query_string = "select * from devices where deviceId in [{}]".format(
-            device_include_string
-        )
-
-        self.command_execute_assert(
-            'iot hub monitor-events -n {} -g {} --device-query "{}" --et {} -t 10 -y -p sys anno app'.format(
-                LIVE_HUB, LIVE_RG, query_string, enqueued_time
-            ),
-            device_subset_include,
-        )
-
-        # Expect failure for excluded devices
-        device_subset_exclude = device_ids[device_count // 2:]
-        with pytest.raises(Exception):
-            self.command_execute_assert(
-                'iot hub monitor-events -n {} -g {} --device-query "{}" --et {} -t 10 -y -p sys anno app'.format(
-                    LIVE_HUB, LIVE_RG, query_string, enqueued_time
-                ),
-                device_subset_exclude,
-            )
-
-        # Monitor events with --login parameter
-        self.command_execute_assert(
-            "iot hub monitor-events -t 10 -y -p all --cg {} --et {} --login {}".format(
-                LIVE_CONSUMER_GROUPS[2], enqueued_time, LIVE_HUB_CS
-            ),
-            device_ids,
-        )
-
-        enqueued_time = calculate_millisec_since_unix_epoch_utc()
-
-        # Send messages that have JSON payload, but do not pass $.ct property
-        execute_onthread(
-            method=iot_device_send_message,
-            args=[
-                client,
-                device_ids[i],
-                LIVE_HUB,
-                '{\r\n"payload_data1":"payload_value1"\r\n}',
-                "",
-                1,
-                LIVE_RG,
-            ],
-            max_runs=1,
-        )
-
-        # Monitor messages for ugly JSON output
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 -y".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[0], enqueued_time
-            ),
-            ["\\r\\n"],
-        )
-
-        # Monitor messages and parse payload as JSON with the --ct parameter
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 --ct application/json -y".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[1], enqueued_time
-            ),
-            ['"payload_data1": "payload_value1"'],
-        )
-
-        enqueued_time = calculate_millisec_since_unix_epoch_utc()
-
-        # Send messages that have JSON payload and have $.ct property
-        execute_onthread(
-            method=iot_device_send_message,
-            args=[
-                client,
-                device_ids[i],
-                LIVE_HUB,
-                '{\r\n"payload_data1":"payload_value1"\r\n}',
-                "$.ct=application/json",
-                1,
-                LIVE_RG,
-            ],
-            max_runs=1,
-        )
-
-        # Monitor messages for pretty JSON output
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 -y".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[0], enqueued_time
-            ),
-            ['"payload_data1": "payload_value1"'],
-        )
-
-        # Monitor messages with yaml output
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 -y -o yaml".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[1], enqueued_time
-            ),
-            ["payload_data1: payload_value1"],
-        )
-
-        enqueued_time = calculate_millisec_since_unix_epoch_utc()
-
-        # Send messages that have improperly formatted JSON payload and a $.ct property
-        execute_onthread(
-            method=iot_device_send_message,
-            args=[
-                client,
-                device_ids[i],
-                LIVE_HUB,
-                '{\r\n"payload_data1""payload_value1"\r\n}',
-                "$.ct=application/json",
-                1,
-                LIVE_RG,
-            ],
-            max_runs=1,
-        )
-
-        # Monitor messages to ensure it returns improperly formatted JSON
-        self.command_execute_assert(
-            "iot hub monitor-events -n {} -g {} --cg {} --et {} -t 10 -y".format(
-                LIVE_HUB, LIVE_RG, LIVE_CONSUMER_GROUPS[0], enqueued_time
-            ),
-            ['{\\r\\n\\"payload_data1\\"\\"payload_value1\\"\\r\\n}'],
-        )
-
-        for cg in LIVE_CONSUMER_GROUPS:
-            self.cmd(
-                "az iot hub consumer-group delete --hub-name {} --resource-group {} --name {}".format(
-                    LIVE_HUB, LIVE_RG, cg
-                ),
-                expect_failure=False,
-            )
-
-    @pytest.mark.skipif(
-        not validate_min_python_version(3, 4, exit_on_fail=False),
-        reason="minimum python version not satisfied",
-    )
-    def test_hub_monitor_feedback(self):
-        device_count = 1
-
-        names = self._create_entity_names(devices=device_count)
-        device_ids = names["device_ids"]
-
-        for i in range(device_count):
-            self.cmd(
-                "iot hub device-identity create -d {} -n {} -g {}".format(
-                    device_ids[i], LIVE_HUB, LIVE_RG
-                ),
-                checks=[self.check("deviceId", device_ids[i])],
-            )
-
-        ack = "full"
-        self.cmd(
-            "iot device c2d-message send -d {} --hub-name {} -g {} --ack {} -y".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, ack
-            ),
-            checks=self.is_empty(),
-        )
-
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --hub-name {} -g {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG
-            )
-        ).get_output_in_json()
-        msg_id = result["messageId"]
-        etag = result["etag"]
-        assert result["ack"] == ack
-
-        self.cmd(
-            "iot device c2d-message complete -d {} --hub-name {} -g {} -e {}".format(
-                device_ids[0], LIVE_HUB, LIVE_RG, etag
-            )
-        )
-
-        self.command_execute_assert(
-            "iot hub monitor-feedback -n {} -g {} -w {} -y".format(
-                LIVE_HUB, LIVE_RG, msg_id
-            ),
-            ["description: Success"],
-        )
-
-        # With connection string - filter on device
-        ack = "positive"
-        self.cmd(
-            "iot device c2d-message send -d {} --login {} --ack {} -y".format(
-                device_ids[0], LIVE_HUB_CS, ack
-            ),
-            checks=self.is_empty(),
-        )
-
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --login {}".format(
-                device_ids[0], LIVE_HUB_CS
-            )
-        ).get_output_in_json()
-        msg_id = result["messageId"]
-        etag = result["etag"]
-        assert result["ack"] == ack
-
-        self.cmd(
-            "iot device c2d-message complete -d {} --login {} -e {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            )
-        )
-
-        self.command_execute_assert(
-            "iot hub monitor-feedback --login {} -w {} -d {} -y".format(
-                LIVE_HUB_CS, msg_id, device_ids[0]
-            ),
-            ["description: Success"],
-        )
-
-        # With connection string - dead lettered case + unrelated ack
-        ack = "negative"
-
-        # Create some noise
-        self.cmd(
-            "iot device c2d-message send -d {} --login {} --ack {} -y".format(
-                device_ids[0], LIVE_HUB_CS, ack
-            ),
-            checks=self.is_empty(),
-        )
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --login {}".format(
-                device_ids[0], LIVE_HUB_CS
-            )
-        ).get_output_in_json()
-        etag = result["etag"]
-        self.cmd(
-            "iot device c2d-message reject -d {} --login {} -e {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            )
-        )
-
-        # Target message
-        self.cmd(
-            "iot device c2d-message send -d {} --login {} --ack {} -y".format(
-                device_ids[0], LIVE_HUB_CS, ack
-            ),
-            checks=self.is_empty(),
-        )
-
-        result = self.cmd(
-            "iot device c2d-message receive -d {} --login {}".format(
-                device_ids[0], LIVE_HUB_CS
-            )
-        ).get_output_in_json()
-        msg_id = result["messageId"]
-        etag = result["etag"]
-        assert result["ack"] == ack
-
-        self.cmd(
-            "iot device c2d-message reject -d {} --login {} -e {}".format(
-                device_ids[0], LIVE_HUB_CS, etag
-            )
-        )
-
-        self.command_execute_assert(
-            "iot hub monitor-feedback --login {} -w {} -y".format(LIVE_HUB_CS, msg_id),
-            ["description: Message rejected"],
+class TestIoTStorage(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTStorage, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
         )
 
     @pytest.mark.skipif(
@@ -2443,8 +1757,7 @@ class TestIoTHub(LiveScenarioTest):
         device_count = 1
 
         content_path = os.path.join(CWD, "test_config_modules_content_v1.json")
-        names = self._create_entity_names(devices=device_count)
-        device_ids = names["device_ids"]
+        device_ids = self.generate_device_names(device_count)
 
         self.cmd(
             "iot hub device-identity create -d {} -n {} -g {} --ee".format(
@@ -2480,23 +1793,27 @@ class TestIoTHub(LiveScenarioTest):
             ],
         )
 
+
+class TestIoTEdgeOffline(IoTLiveScenarioTest):
+    def __init__(self, test_case):
+        super(TestIoTEdgeOffline, self).__init__(
+            test_case, LIVE_HUB, LIVE_RG, LIVE_HUB_CS
+        )
+
     def test_edge_offline(self):
         device_count = 3
         edge_device_count = 2
 
-        names = self._create_entity_names(
-            devices=device_count, edge_devices=edge_device_count
-        )
-        device_ids = names["device_ids"]
-        edge_device_ids = names["edge_device_ids"]
+        device_ids = self.generate_device_names(device_count)
+        edge_device_ids = self.generate_device_names(edge_device_count, True)
 
-        for i in range(edge_device_count):
+        for edge_device in edge_device_ids:
             self.cmd(
                 "iot hub device-identity create -d {} -n {} -g {} --ee".format(
-                    edge_device_ids[i], LIVE_HUB, LIVE_RG
+                    edge_device, LIVE_HUB, LIVE_RG
                 ),
                 checks=[
-                    self.check("deviceId", edge_device_ids[i]),
+                    self.check("deviceId", edge_device),
                     self.check("status", "enabled"),
                     self.check("statusReason", None),
                     self.check("connectionState", "Disconnected"),
@@ -2507,13 +1824,13 @@ class TestIoTHub(LiveScenarioTest):
                 ],
             )
 
-        for i in range(device_count):
+        for device in device_ids:
             self.cmd(
                 "iot hub device-identity create -d {} -n {} -g {}".format(
-                    device_ids[i], LIVE_HUB, LIVE_RG
+                    device, LIVE_HUB, LIVE_RG
                 ),
                 checks=[
-                    self.check("deviceId", device_ids[i]),
+                    self.check("deviceId", device),
                     self.check("status", "enabled"),
                     self.check("statusReason", None),
                     self.check("connectionState", "Disconnected"),
@@ -2630,6 +1947,8 @@ class TestIoTHub(LiveScenarioTest):
             ),
             expect_failure=False,
         )
+
+        # TODO: Result should be JSON
         expected_output = "{}".format(device_ids[1])
         assert output.get_output_in_json() == expected_output
