+++ azext_iot/tests/test_iot_pnp_unit.py
@@ -13,8 +13,7 @@ from azext_iot.operations import pnp as subject
 from azext_iot.common.utility import url_encode_str
 from knack.util import CLIError
 from azure.cli.core.util import read_file_content
-from .test_iot_ext_unit import fixture_cmd, path_service_client, build_mock_response
-
+from .conftest import fixture_cmd, path_service_client, build_mock_response
 
 _repo_endpoint = "https://{}.{}".format(str(uuid4()), "com")
 _repo_id = str(uuid4()).replace("-", "")
