+++ dev_requirements
@@ -6,3 +6,4 @@ pylint
 flake8
 black;python_version>='3.6'
 wheel==0.30.0
+attrs==17.4.0
