+++ setup.cfg
@@ -11,3 +11,4 @@ ignore =
     C901,  # code flow is too complex, to be removed in the future
     W503   # line break after binary operator
     W504   # line break after binary operator
+    E203   # whitespace before ':'
